Docker integrated

+++ contracts/adapters/HyperSwapV2Adapter.sol
@@ -7,18 +7,20 @@ import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
 interface IHyperSwapV2Router {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] calldata path)
-        external view returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] calldata path)
+        external
+        view
+        returns (uint256[] memory amounts);
 
     function factory() external pure returns (address);
-    
+
     function WETH() external pure returns (address);
 }
 
@@ -34,26 +36,24 @@ interface IHyperSwapV2Pair {
 
 contract HyperSwapV2Adapter is IDEXAdapter {
     using SafeERC20 for IERC20;
-    
+
     IHyperSwapV2Router public immutable router;
     IHyperSwapV2Factory public immutable factory;
-    
+
     uint256 private constant FEE_RATE = 300; // 0.3%
-    
+
     constructor(address _router) {
         router = IHyperSwapV2Router(_router);
         factory = IHyperSwapV2Factory(router.factory());
     }
 
-    function swap(
-        address tokenIn,
-        address tokenOut,
-        uint256 amountIn,
-        uint256 minAmountOut,
-        bytes calldata swapData
-    ) external override returns (uint256 amountOut) {
+    function swap(address tokenIn, address tokenOut, uint256 amountIn, uint256 minAmountOut, bytes calldata swapData)
+        external
+        override
+        returns (uint256 amountOut)
+    {
         address[] memory path;
-        
+
         if (swapData.length > 0) {
             path = abi.decode(swapData, (address[]));
         } else {
@@ -61,31 +61,26 @@ contract HyperSwapV2Adapter is IDEXAdapter {
             path[0] = tokenIn;
             path[1] = tokenOut;
         }
-        
+
         require(path[0] == tokenIn && path[path.length - 1] == tokenOut, "HyperSwapV2: Invalid path");
 
         IERC20(tokenIn).safeTransferFrom(msg.sender, address(this), amountIn);
         IERC20(tokenIn).safeApprove(address(router), amountIn);
 
-        uint[] memory amounts = router.swapExactTokensForTokens(
-            amountIn,
-            minAmountOut,
-            path,
-            msg.sender,
-            block.timestamp + 300
-        );
+        uint256[] memory amounts =
+            router.swapExactTokensForTokens(amountIn, minAmountOut, path, msg.sender, block.timestamp + 300);
 
         return amounts[amounts.length - 1];
     }
 
-    function getQuote(
-        address tokenIn,
-        address tokenOut,
-        uint256 amountIn,
-        bytes calldata swapData
-    ) external view override returns (uint256 amountOut, uint256 gasEstimate) {
+    function getQuote(address tokenIn, address tokenOut, uint256 amountIn, bytes calldata swapData)
+        external
+        view
+        override
+        returns (uint256 amountOut, uint256 gasEstimate)
+    {
         address[] memory path;
-        
+
         if (swapData.length > 0) {
             path = abi.decode(swapData, (address[]));
         } else {
@@ -93,8 +88,8 @@ contract HyperSwapV2Adapter is IDEXAdapter {
             path[0] = tokenIn;
             path[1] = tokenOut;
         }
-        
-        try router.getAmountsOut(amountIn, path) returns (uint[] memory amounts) {
+
+        try router.getAmountsOut(amountIn, path) returns (uint256[] memory amounts) {
             return (amounts[amounts.length - 1], 150000);
         } catch {
             return (0, 150000);
@@ -104,7 +99,7 @@ contract HyperSwapV2Adapter is IDEXAdapter {
     function getDEXName() external pure override returns (string memory) {
         return "HyperSwap V2";
     }
-    
+
     function getSupportedFeeTiers() external pure override returns (uint256[] memory) {
         uint256[] memory fees = new uint256[](1);
         fees[0] = FEE_RATE;
@@ -114,10 +109,10 @@ contract HyperSwapV2Adapter is IDEXAdapter {
     function getPairLiquidity(address tokenA, address tokenB) external view returns (uint256 liquidity) {
         address pair = factory.getPair(tokenA, tokenB);
         if (pair == address(0)) return 0;
-        
+
         (uint112 reserve0, uint112 reserve1,) = IHyperSwapV2Pair(pair).getReserves();
         address token0 = IHyperSwapV2Pair(pair).token0();
-        
+
         if (token0 == tokenA) {
             liquidity = uint256(reserve0);
         } else {
@@ -125,21 +120,21 @@ contract HyperSwapV2Adapter is IDEXAdapter {
         }
     }
 
-    function calculatePriceImpact(
-        address tokenIn,
-        address tokenOut,
-        uint256 amountIn
-    ) external view returns (uint256 priceImpact) {
+    function calculatePriceImpact(address tokenIn, address tokenOut, uint256 amountIn)
+        external
+        view
+        returns (uint256 priceImpact)
+    {
         address pair = factory.getPair(tokenIn, tokenOut);
         if (pair == address(0)) return 10000; // 100% if no pair
-        
+
         (uint112 reserve0, uint112 reserve1,) = IHyperSwapV2Pair(pair).getReserves();
         address token0 = IHyperSwapV2Pair(pair).token0();
-        
+
         uint256 reserveIn = token0 == tokenIn ? uint256(reserve0) : uint256(reserve1);
-        
+
         if (reserveIn == 0) return 10000;
-        
+
         //Price impact = (amountIn / reserveIn) * 10000
         priceImpact = (amountIn * 10000) / reserveIn;
         return priceImpact > 10000 ? 10000 : priceImpact;
