+++ contracts/interfaces/IKittenSwap.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.19;
 
-
 interface IKittenSwapRouter {
     function factory() external pure returns (address);
     function WETH() external pure returns (address);
@@ -10,46 +9,50 @@ interface IKittenSwapRouter {
         address tokenA,
         address tokenB,
         bool stable,
-        uint amountADesired,
-        uint amountBDesired,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB, uint liquidity);
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB, uint256 liquidity);
 
     function removeLiquidity(
         address tokenA,
         address tokenB,
         bool stable,
-        uint liquidity,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 liquidity,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB);
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB);
 
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
     function swapExactTokensForTokensStable(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] calldata path)
-        external view returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] calldata path)
+        external
+        view
+        returns (uint256[] memory amounts);
 
-    function getAmountsOutStable(uint amountIn, address[] calldata path)
-        external view returns (uint[] memory amounts);
+    function getAmountsOutStable(uint256 amountIn, address[] calldata path)
+        external
+        view
+        returns (uint256[] memory amounts);
 
     // ve(3,3) specific functions
     function getRewards(address account, address[] calldata tokens) external view returns (uint256[] memory);
@@ -59,13 +62,13 @@ interface IKittenSwapRouter {
 }
 
 interface IKittenSwapFactory {
-    event PairCreated(address indexed token0, address indexed token1, bool stable, address pair, uint);
+    event PairCreated(address indexed token0, address indexed token1, bool stable, address pair, uint256);
 
     function isPair(address pair) external view returns (bool);
     function isStable(address pair) external view returns (bool);
     function getPair(address tokenA, address tokenB, bool stable) external view returns (address);
-    function allPairs(uint) external view returns (address);
-    function allPairsLength() external view returns (uint);
+    function allPairs(uint256) external view returns (address);
+    function allPairsLength() external view returns (uint256);
 
     function createPair(address tokenA, address tokenB, bool stable) external returns (address pair);
     function getInitializable() external view returns (address, address, bool);
@@ -77,16 +80,16 @@ interface IKittenSwapFactory {
 }
 
 interface IKittenSwapPair {
-    event Approval(address indexed owner, address indexed spender, uint value);
-    event Transfer(address indexed from, address indexed to, uint value);
-    event Mint(address indexed sender, uint amount0, uint amount1);
-    event Burn(address indexed sender, uint amount0, uint amount1, address indexed to);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Mint(address indexed sender, uint256 amount0, uint256 amount1);
+    event Burn(address indexed sender, uint256 amount0, uint256 amount1, address indexed to);
     event Swap(
         address indexed sender,
-        uint amount0In,
-        uint amount1In,
-        uint amount0Out,
-        uint amount1Out,
+        uint256 amount0In,
+        uint256 amount1In,
+        uint256 amount0Out,
+        uint256 amount1Out,
         address indexed to
     );
     event Sync(uint112 reserve0, uint112 reserve1);
@@ -94,33 +97,34 @@ interface IKittenSwapPair {
     function name() external pure returns (string memory);
     function symbol() external pure returns (string memory);
     function decimals() external pure returns (uint8);
-    function totalSupply() external view returns (uint);
-    function balanceOf(address owner) external view returns (uint);
-    function allowance(address owner, address spender) external view returns (uint);
+    function totalSupply() external view returns (uint256);
+    function balanceOf(address owner) external view returns (uint256);
+    function allowance(address owner, address spender) external view returns (uint256);
 
-    function approve(address spender, uint value) external returns (bool);
-    function transfer(address to, uint value) external returns (bool);
-    function transferFrom(address from, address to, uint value) external returns (bool);
+    function approve(address spender, uint256 value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
+    function transferFrom(address from, address to, uint256 value) external returns (bool);
 
     function DOMAIN_SEPARATOR() external view returns (bytes32);
     function PERMIT_TYPEHASH() external pure returns (bytes32);
-    function nonces(address owner) external view returns (uint);
+    function nonces(address owner) external view returns (uint256);
 
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s)
+        external;
 
-    function MINIMUM_LIQUIDITY() external pure returns (uint);
+    function MINIMUM_LIQUIDITY() external pure returns (uint256);
     function factory() external view returns (address);
     function token0() external view returns (address);
     function token1() external view returns (address);
     function stable() external view returns (bool);
     function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function price0CumulativeLast() external view returns (uint);
-    function price1CumulativeLast() external view returns (uint);
-    function kLast() external view returns (uint);
+    function price0CumulativeLast() external view returns (uint256);
+    function price1CumulativeLast() external view returns (uint256);
+    function kLast() external view returns (uint256);
 
-    function mint(address to) external returns (uint liquidity);
-    function burn(address to) external returns (uint amount0, uint amount1);
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+    function mint(address to) external returns (uint256 liquidity);
+    function burn(address to) external returns (uint256 amount0, uint256 amount1);
+    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data) external;
     function skim(address to) external;
     function sync() external;
 
@@ -128,25 +132,25 @@ interface IKittenSwapPair {
 }
 
 interface IKittenSwapGauge {
-    function deposit(uint amount, uint tokenId) external;
-    function depositAll(uint tokenId) external;
-    function withdraw(uint amount) external;
+    function deposit(uint256 amount, uint256 tokenId) external;
+    function depositAll(uint256 tokenId) external;
+    function withdraw(uint256 amount) external;
     function withdrawAll() external;
-    function withdrawToken(uint amount, uint tokenId) external;
+    function withdrawToken(uint256 amount, uint256 tokenId) external;
 
     function getReward(address account, address[] memory tokens) external;
-    function claimFees() external returns (uint claimed0, uint claimed1);
+    function claimFees() external returns (uint256 claimed0, uint256 claimed1);
 
-    function left(address token) external view returns (uint);
+    function left(address token) external view returns (uint256);
     function isForPair() external view returns (bool);
     function stakingToken() external view returns (address);
     function rewardToken() external view returns (address);
-    function earned(address token, address account) external view returns (uint);
-    function totalSupply() external view returns (uint);
-    function balanceOf(address account) external view returns (uint);
-    function rewardRate(address token) external view returns (uint);
-    function rewardPerToken(address token) external view returns (uint);
-    function lastTimeRewardApplicable(address token) external view returns (uint);
+    function earned(address token, address account) external view returns (uint256);
+    function totalSupply() external view returns (uint256);
+    function balanceOf(address account) external view returns (uint256);
+    function rewardRate(address token) external view returns (uint256);
+    function rewardPerToken(address token) external view returns (uint256);
+    function lastTimeRewardApplicable(address token) external view returns (uint256);
 }
 
 interface IVotingEscrow {
@@ -157,35 +161,35 @@ interface IVotingEscrow {
         uint256 blk;
     }
 
-    function create_lock_for(uint _value, uint _lock_duration, address _to) external returns (uint);
-    function create_lock(uint _value, uint _lock_duration) external returns (uint);
-    function increase_amount(uint _tokenId, uint _value) external;
-    function increase_unlock_time(uint _tokenId, uint _lock_duration) external;
-    function withdraw(uint _tokenId) external;
-    function merge(uint _from, uint _to) external;
-    function split(uint _tokenId, uint _amount) external returns (uint);
+    function create_lock_for(uint256 _value, uint256 _lock_duration, address _to) external returns (uint256);
+    function create_lock(uint256 _value, uint256 _lock_duration) external returns (uint256);
+    function increase_amount(uint256 _tokenId, uint256 _value) external;
+    function increase_unlock_time(uint256 _tokenId, uint256 _lock_duration) external;
+    function withdraw(uint256 _tokenId) external;
+    function merge(uint256 _from, uint256 _to) external;
+    function split(uint256 _tokenId, uint256 _amount) external returns (uint256);
 
-    function locked(uint _tokenId) external view returns (uint amount, uint end);
-    function balanceOfNFT(uint _tokenId) external view returns (uint);
-    function totalSupply() external view returns (uint);
-    function supply() external view returns (uint);
+    function locked(uint256 _tokenId) external view returns (uint256 amount, uint256 end);
+    function balanceOfNFT(uint256 _tokenId) external view returns (uint256);
+    function totalSupply() external view returns (uint256);
+    function supply() external view returns (uint256);
 
-    function ownerOf(uint _tokenId) external view returns (address);
-    function balanceOf(address _owner) external view returns (uint);
-    function tokenOfOwnerByIndex(address _owner, uint _tokenIndex) external view returns (uint);
-    function isApprovedOrOwner(address _spender, uint _tokenId) external view returns (bool);
+    function ownerOf(uint256 _tokenId) external view returns (address);
+    function balanceOf(address _owner) external view returns (uint256);
+    function tokenOfOwnerByIndex(address _owner, uint256 _tokenIndex) external view returns (uint256);
+    function isApprovedOrOwner(address _spender, uint256 _tokenId) external view returns (bool);
 
-    function user_point_epoch(uint _tokenId) external view returns (uint);
-    function epoch() external view returns (uint);
-    function user_point_history(uint _tokenId, uint _loc) external view returns (Point memory);
-    function point_history(uint _loc) external view returns (Point memory);
+    function user_point_epoch(uint256 _tokenId) external view returns (uint256);
+    function epoch() external view returns (uint256);
+    function user_point_history(uint256 _tokenId, uint256 _loc) external view returns (Point memory);
+    function point_history(uint256 _loc) external view returns (Point memory);
     function checkpoint() external;
-    function deposit_for(uint _tokenId, uint _value) external;
+    function deposit_for(uint256 _tokenId, uint256 _value) external;
 
-    function voted(uint _tokenId) external view returns (bool);
-    function attachments(uint _tokenId) external view returns (uint);
-    function voting(uint _tokenId) external;
-    function abstain(uint _tokenId) external;
-    function attach(uint _tokenId) external;
-    function detach(uint _tokenId) external;
+    function voted(uint256 _tokenId) external view returns (bool);
+    function attachments(uint256 _tokenId) external view returns (uint256);
+    function voting(uint256 _tokenId) external;
+    function abstain(uint256 _tokenId) external;
+    function attach(uint256 _tokenId) external;
+    function detach(uint256 _tokenId) external;
 }
