+++ tests/contracts/mocks/MockDEXRouter.sol
@@ -9,43 +9,49 @@ contract MockDEXRouter {
 
     // Mock HyperSwap V2 Router
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint /*amountOutMin*/,
+        uint256 amountIn,
+        uint256, /*amountOutMin*/
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts) {
+        uint256 deadline
+    ) external returns (uint256[] memory amounts) {
         require(path.length >= 2, "Invalid path");
         require(block.timestamp <= deadline, "Expired");
-        
-        amounts = new uint[](path.length);
+
+        amounts = new uint256[](path.length);
         amounts[0] = amountIn;
-        
+
         // Transfer input tokens
         IERC20(path[0]).safeTransferFrom(msg.sender, address(this), amountIn);
-        
+
         // Calculate output (simulate 0.3% fee)
         uint256 amountOut = (amountIn * 997) / 1000;
         amounts[path.length - 1] = amountOut;
-        
+
         // Transfer output tokens
         IERC20(path[path.length - 1]).safeTransfer(to, amountOut);
     }
 
-    function getAmountsOut(uint amountIn, address[] calldata path)
-        external pure returns (uint[] memory amounts) {
-        amounts = new uint[](path.length);
+    function getAmountsOut(uint256 amountIn, address[] calldata path)
+        external
+        pure
+        returns (uint256[] memory amounts)
+    {
+        amounts = new uint256[](path.length);
         amounts[0] = amountIn;
-        
+
         // Simulate output calculation
         amounts[path.length - 1] = (amountIn * 997) / 1000;
     }
 
-    function getAmountsOutStable(uint amountIn, address[] calldata path)
-        external pure returns (uint[] memory amounts) {
-        amounts = new uint[](path.length);
+    function getAmountsOutStable(uint256 amountIn, address[] calldata path)
+        external
+        pure
+        returns (uint256[] memory amounts)
+    {
+        amounts = new uint256[](path.length);
         amounts[0] = amountIn;
-        
+
         // Simulate stable swap with lower fees
         amounts[path.length - 1] = (amountIn * 9999) / 10000;
     }
@@ -62,27 +68,26 @@ contract MockDEXRouter {
         uint160 sqrtPriceLimitX96;
     }
 
-    function exactInputSingle(ExactInputSingleParams calldata params)
-        external returns (uint256 amountOut) {
+    function exactInputSingle(ExactInputSingleParams calldata params) external returns (uint256 amountOut) {
         require(block.timestamp <= params.deadline, "Expired");
-        
+
         // Transfer input tokens
         IERC20(params.tokenIn).safeTransferFrom(msg.sender, address(this), params.amountIn);
-        
+
         // Calculate output based on fee tier
         uint256 feeRate = params.fee;
         amountOut = (params.amountIn * (1000000 - feeRate)) / 1000000;
-        
+
         require(amountOut >= params.amountOutMinimum, "Insufficient output");
-        
+
         // Transfer output tokens
         IERC20(params.tokenOut).safeTransfer(params.recipient, amountOut);
     }
 
     // Mock quoter
     function quoteExactInputSingle(
-        address /*tokenIn*/,
-        address /*tokenOut*/,
+        address, /*tokenIn*/
+        address, /*tokenOut*/
         uint24 fee,
         uint256 amountIn,
         uint160 /*sqrtPriceLimitX96*/
@@ -91,11 +96,11 @@ contract MockDEXRouter {
     }
 
     // Mock Laminar Router
-    function getOptimalSwap(
-        address tokenIn,
-        address tokenOut,
-        uint256 amountIn
-    ) external pure returns (uint256 amountOut, bytes memory routeData) {
+    function getOptimalSwap(address tokenIn, address tokenOut, uint256 amountIn)
+        external
+        pure
+        returns (uint256 amountOut, bytes memory routeData)
+    {
         amountOut = (amountIn * 9998) / 10000; // 0.02% fee
         routeData = abi.encode("laminar_route");
     }
@@ -109,10 +114,10 @@ contract MockDEXRouter {
     ) external returns (uint256 amountOut) {
         // Transfer input tokens
         IERC20(tokenIn).safeTransferFrom(msg.sender, address(this), amountIn);
-        
+
         amountOut = (amountIn * 9998) / 10000;
         require(amountOut >= minAmountOut, "Insufficient output");
-        
+
         // Transfer output tokens
         IERC20(tokenOut).safeTransfer(msg.sender, amountOut);
     }
